+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "workout-scheduler",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "expo-router/entry",
   "scripts": {
     "ts:check": "tsc",
     "start": "expo start",
